requirements.txt
streamlit
requests
duckdb
Pillow
toml
langchain_nvidia_ai_endpoints
folium